inch display add



See merge request !249

Former-commit-id: 7a2574b454f01b95ca171150b31805c2a7b78bf8

+++ dev/js/services/design_serv.js
@@ -398,7 +398,7 @@
                     var deff = $q.defer();
                     if (DesignStor.design.tempSize.length) {
                         if (UserStor.userInfo.factory_id === 897) {
-                            var newLength = parseInt(DesignStor.design.tempSize.join(''), 10) / 0.0393701;
+                            var newLength = parseFloat(DesignStor.design.tempSize.join(''), 10) / 0.0393701;
                         } else {
                             var newLength = parseInt(DesignStor.design.tempSize.join(''), 10);
                         }
